# server/swagger-debug.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gold Pawn Broker API",
    "version": "1.0.0",
    "description": "API for managing gold pawn broker operations including particulars, transactions, and dashboard analytics",
    "contact": {
      "name": "API Support",
      "email": "support@goldpawnbroker.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
...
  "paths": {
...
    "/api/transactions/particular/{particularId}": {
      "get": {
        "summary": "Get all transactions for a particular",
        "tags": [
          "Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "particularId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "transactionType",
            "schema": {
              "type": "string",
              "enum": [
                "cash",
                "metal"
              ]
            }
          },
          {
            "in": "query",
            "name": "transactionFlow",
            "schema": {
              "type": "string",
              "enum": [
                "incoming",
                "outgoing"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of transactions"
          }
        }
      }
    },
    "/api/transactions": {
      "post": {
        "summary": "Create a new transaction",
        "tags": [
          "Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "particularId",
                  "transactionType",
                  "transactionFlow",
                  "quantity",
                  "total"
                ],
                "properties": {
                  "particularId": {
                    "type": "string"
                  },
                  "transactionType": {
                    "type": "string",
                    "enum": [
                      "cash",
                      "metal"
                    ]
                  },
                  "transactionFlow": {
                    "type": "string",
                    "enum": [
                      "incoming",
                      "outgoing"
                    ]
                  },
                  "quantity": {
                    "type": "number"
                  },
                  "rate": {
                    "type": "number"
                  },
                  "percentage": {
                    "type": "number"
                  },
                  "total": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully"
          }
        }
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "summary": "Get a specific transaction",
        "tags": [
          "Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction details"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      },
      "put": {
        "summary": "Update a transaction",
        "tags": [
          "Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "number"
                  },
                  "rate": {
                    "type": "number"
                  },
                  "percentage": {
                    "type": "number"
                  },
                  "total": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction updated successfully"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a transaction",
        "tags": [
          "Transactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction deleted successfully"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      }
    }
  }
}